Uploading refactoring: wip
