+++ lib/View/templates/_upload.html
@@ -594,15 +594,15 @@
   var clientId = '${clientId}';
 
 
-  window.onGDriveApiLoad = function () {
-    setTimeout(()=> {
-      if (window.onGDriveApiInit) {
-        window.onGDriveApiInit()
-      } else {
-        window.onGDriveApiLoad
-      }
-    }, 500)
-  }
+  // window.onGDriveApiLoad = function () {
+  //   setTimeout(()=> {
+  //     if (window.onGDriveApiInit) {
+  //       window.onGDriveApiInit()
+  //     } else {
+  //       window.onGDriveApiLoad
+  //     }
+  //   }, 500)
+  // }
 
 </script>
 <script tal:attributes="nonce x_nonce_unique_id" type="text/javascript" src="//apis.google.com/js/api.js?onload=onGDriveApiLoad"></script>
