+++ public/js/cat_source/es6/components/createProject/UploadGdrive.js
@@ -1,0 +1,237 @@
+import React, {useEffect, useState} from 'react'
+import UserStore from '../../stores/UserStore'
+import ModalsActions from '../../actions/ModalsActions'
+import {getUserConnectedService} from '../../api/getUserConnectedService'
+import CreateProjectStore from '../../stores/CreateProjectStore'
+import {openGDriveFiles} from '../../api/openGDriveFiles'
+import CreateProjectActions from '../../actions/CreateProjectActions'
+import {getGoogleDriveUploadedFiles} from '../../api/getGoogleDriveUploadedFiles'
+import CommonUtils from '../../utils/commonUtils'
+import {getPrintableFileSize} from './UploadFile'
+import {Button, BUTTON_SIZE, BUTTON_TYPE} from '../common/Button/Button'
+import {DeleteIcon} from '../segments/SegmentFooterTabGlossary'
+import {deleteGDriveUploadedFile} from '../../api/deleteGdriveUploadedFile'
+import IconClose from '../icons/IconClose'
+
+export const UploadGdrive = ({
+  openGdrive,
+  uploadedFilesNames,
+  sourceLang,
+  targetLangs,
+  xliffConfigTemplateId,
+  segmentationRule,
+  setUploadFilesType,
+}) => {
+  const [authApiLoaded, setAuthApiLoaded] = useState(false)
+  const [pickerApiLoaded, setPickerApiLoaded] = useState(false)
+  const [loading, setLoading] = useState(false)
+  const [files, setFiles] = useState([])
+
+  useEffect(() => {
+    gapi.load('auth', {callback: setAuthApiLoaded(true)})
+    gapi.load('picker', {callback: setPickerApiLoaded(true)})
+  }, [])
+
+  useEffect(() => {
+    openGdrive && openGdrivePicker()
+  }, [openGdrive])
+
+  const gdriveInitComplete = () => {
+    return pickerApiLoaded && authApiLoaded
+  }
+
+  const tryToRefreshToken = (service) => {
+    return getUserConnectedService(service.id)
+  }
+
+  const openGdrivePicker = () => {
+    if (!gdriveInitComplete()) {
+      console.log('gdriveInitComplete not complete')
+      return
+    }
+
+    const defaultService = UserStore.getDefaultConnectedService()
+
+    if (!defaultService) {
+      showPreferencesWithMessage()
+      return
+    }
+
+    tryToRefreshToken(defaultService)
+      .then((data) => {
+        UserStore.updateConnectedService(data.connected_service)
+        createPicker(UserStore.getDefaultConnectedService())
+      })
+      .catch(() => {
+        UserStore.updateConnectedService({defaultService, is_default: false})
+        showPreferencesWithMessage()
+      })
+  }
+
+  const createPicker = (service) => {
+    const token = JSON.parse(service.oauth_access_token)
+
+    const picker = new google.picker.PickerBuilder()
+      .setAppId(window.clientId)
+      .addView(google.picker.ViewId.DOCUMENTS)
+      .addView(google.picker.ViewId.PRESENTATIONS)
+      .addView(google.picker.ViewId.SPREADSHEETS)
+      .setOAuthToken(token.access_token)
+      .setDeveloperKey(window.developerKey)
+      .setCallback(pickerCallback)
+      .enableFeature(google.picker.Feature.MINE_ONLY)
+      .enableFeature(google.picker.Feature.MULTISELECT_ENABLED)
+      .build()
+    try {
+      picker.setVisible(true)
+    } catch (e) {
+      UserStore.updateConnectedService({service, is_default: false})
+      throw new Error('Picker Error')
+    }
+  }
+
+  const pickerCallback = (data) => {
+    if (data[google.picker.Response.ACTION] == google.picker.Action.PICKED) {
+      let exportIds = []
+      data[google.picker.Response.DOCUMENTS].forEach((doc) => {
+        exportIds.push(doc.id)
+      })
+
+      const jsonDoc = {
+        exportIds: exportIds,
+        action: 'open',
+      }
+
+      setLoading(true)
+      openGDriveFiles({
+        encodedJson: encodeURIComponent(JSON.stringify(jsonDoc)),
+        sourceLang: sourceLang,
+        targetLang: targetLangs.map((lang) => lang.id).join(),
+        segmentation_rule: segmentationRule,
+        filters_extraction_parameters_template_id: xliffConfigTemplateId,
+      }).then((response) => {
+        CreateProjectActions.hideErrors()
+        if (response.success) {
+          tryListGDriveFiles()
+        } else {
+          let message =
+            'There was an error retrieving the file from Google Drive. Try again and if the error persists contact the Support.'
+          if (response.error_class === 'Google\\Service\\Exception') {
+            message =
+              'There was an error retrieving the file from Google Drive: ' +
+              response.error_msg
+          }
+          if (response.error_code === 404) {
+            message = (
+              <span>
+                File retrieval error. To find out how to translate the desired
+                file, please{' '}
+                <a
+                  href="https://guides.matecat.com/google-drive-files-upload-issues"
+                  target="_blank"
+                >
+                  read this guide
+                </a>
+                .
+              </span>
+            )
+          }
+
+          CreateProjectActions.showError(message)
+
+          console.error(
+            'Error when processing request. Error class: ' +
+              response.error_class +
+              ', Error code: ' +
+              response.error_code +
+              ', Error message: ' +
+              message,
+          )
+        }
+        setLoading(false)
+      })
+    }
+  }
+
+  const deleteGDriveFile = (file) => {
+    deleteGDriveUploadedFile(file.id).then((response) => {
+      uploadedFilesNames = uploadedFilesNames.filter((f) => f !== file.name)
+      if (response.success) {
+        tryListGDriveFiles()
+      }
+    })
+  }
+
+  const tryListGDriveFiles = () => {
+    getGoogleDriveUploadedFiles().then((listFiles) => {
+      let filesList = []
+      if (listFiles && listFiles.files) {
+        listFiles.files.forEach((file) => {
+          uploadedFilesNames.push(file.fileName)
+          filesList.push({
+            name: file.fileName,
+            ext: file.fileExtension,
+            size: file.fileSize,
+            id: file.fileId,
+          })
+        })
+      }
+      setFiles(filesList)
+    })
+  }
+
+  const showPreferencesWithMessage = () => {
+    ModalsActions.openPreferencesModal({showGDriveMessage: true})
+  }
+
+  return (
+    <div
+      className={`upload-files-container ${files.length > 0 ? 'add-files' : ''}`}
+    >
+      {loading && <div className="fileupload-loading" />}
+      {files.length > 0 && (
+        <>
+          <div className="upload-files-list">
+            {files.map((f, idx) => (
+              <div key={idx} className="file-item">
+                <div className="file-item-name">
+                  <span
+                    className={`file-icon ${CommonUtils.getIconClass(f.ext)}`}
+                  />
+                  {f.name}
+                </div>
+                <div>{getPrintableFileSize(f.size)}</div>
+                <Button
+                  size={BUTTON_SIZE.ICON_SMALL}
+                  style={{marginLeft: 'auto'}}
+                  tooltip={'Remove file'}
+                  onClick={() => deleteGDriveFile(f)}
+                >
+                  <DeleteIcon />
+                </Button>
+              </div>
+            ))}
+          </div>
+          <div className="upload-files-buttons">
+            <Button
+              type={BUTTON_TYPE.PRIMARY}
+              onClick={() => openGdrivePicker()}
+            >
+              <img
+                src="/public/img/logo-drive-16.png"
+                alt="Google drive logo"
+              />
+              Add from Google Drive
+            </Button>
+            <Button
+              type={BUTTON_TYPE.WARNING}
+              onClick={() => files.forEach((f) => deleteGDriveFile(f))}
+            >
+              <IconClose /> Clear all
+            </Button>
+          </div>
+        </>
+      )}
+    </div>
+  )
+}
